requirements.txt
certifi==2017.4.17
chardet==3.0.4
coverage==4.4.1
coveralls==1.1
Django==1.11.4
django-cors-headers==2.1.0
django-nose==1.4.4
djangorestframework==3.6.3
docopt==0.6.2
idna==2.5
nose==1.3.7
pip==9.0.1
pytz==2017.2
requests==2.18.1
setuptools==36.2.2
urllib3==1.21.1
django-filter==1.0.4
django-rest-swagger==2.1.2